Fix sonar cloud issues
